refactor: merge sim and data together.
