+++ src/mrinufft/extras/__init__.py
@@ -1,11 +1,8 @@
 """Additional supports routines."""
 
-from .data import get_brainweb_map
+from .data import get_brainweb_map, fse_simulation
 from .field_map import make_b0map, make_t2smap
-from .sim import fse_simulation
-from .smaps import low_frequency
-from .utils import get_smaps
-
+from .smaps import low_frequency, get_smaps
 
 __all__ = [
     "fse_simulation",
